Undo/redo + major bugfixes
Added undo/redo feature + everything works perfectly now: No flickering
of the mouse. No need to select the canvas before interacting with
elements on it. Added a moveShapeCommand to allow for undo/redo when
moving classes. Window is now resizable (only canvas is still not
adjusting). General cleanup and bugfixing.